LayoutTests/media/now-playing-info-media-session-private-browsing-expected.txt


Tests that the NowPlaying metadata is empty when in private browsing mode.

* NowPlaying should not be active before playback has started.
RUN(video.src = findMediaFile("video", "content/test"))
EVENT(canplaythrough)
RUN(nowPlayingState = internals.nowPlayingState)
EXPECTED (nowPlayingState.registeredAsNowPlayingApplication == 'false') OK
RUN(navigator.mediaSession.metadata = new MediaMetadata({title: "title", artist: "artist", album: "album", artwork: [{src: "content/abe.png"}]}))
RUN(navigator.mediaSession.playbackState = "paused")
RUN(navigator.mediaSession.positionState = {duration: video.duration, position: video.currentTime})
RUN(navigator.mediaSession.setActionHandler("play", playAction))
RUN(navigator.mediaSession.setActionHandler("pause", pauseAction))

* Start to play, NowPlaying should become active.
RUN(navigator.mediaSession.callActionHandler({action: "play"}))
ACTION: play
EVENT(playing)
RUN(navigator.mediaSession.callActionHandler({action: "pause"}))
ACTION: pause

* Metadata should be blank due to private browsing mode.
EXPECTED (internals.nowPlayingMetadata.title == 'undefined') OK
EXPECTED (internals.nowPlayingMetadata.artist == 'undefined') OK
EXPECTED (internals.nowPlayingMetadata.album == 'undefined') OK
EXPECTED (internals.nowPlayingMetadata.artwork == 'undefined') OK

END OF TEST
